+++ lib/state/getit_state_holder.dart
@@ -1,9 +1,9 @@
 import 'package:get_it/get_it.dart';
 import 'package:ministate/state/state_holder.dart';
 
-class GetItStateHolder<STATE, SERVICE extends Object>
+class DefaultStateHolder<STATE, SERVICE extends Object>
     extends StateHolder<STATE, SERVICE> {
-  GetItStateHolder(STATE value) : super(value);
+  DefaultStateHolder(STATE value) : super(value);
 
   @override
   void setState(STATE value) {
@@ -16,5 +16,10 @@ class GetItStateHolder<STATE, SERVICE extends Object>
     // locator Pattern. This assumes a service of type SERVICE
     // is registered in Getit somewhere
     return GetIt.instance<SERVICE>();
+  }
+
+  @override
+  STATE getState() {
+    return value;
   }
 }