+++ lib/state/state_holder.dart
@@ -4,5 +4,6 @@ abstract class StateHolder<STATE, SERVICE extends Object>
     extends ValueNotifier<STATE> {
   StateHolder(STATE value) : super(value);
   void setState(STATE value);
+  STATE getState();
   SERVICE getService();
 }