+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
+A very minimalistic approach to State Management in Flutter inspired by 
+[State Management for Minimalists](https://suragch.medium.com/flutter-state-management-for-minimalists-4c71a2f2f0c1)
 
-* TODO: Describe initial release.
+This first release contains 
+
+* StateHolder abstraction holding logic
+* MiniStateBuilder Widget